+++ fractal_server/migrations/versions/091b01f51f88_add_usergroup_and_linkusergroup_table.py
@@ -7,7 +7,6 @@ Create Date: 2024-09-09 13:17:51.008231
 """
 
 import sqlalchemy as sa
-import sqlmodel.sql.sqltypes
 from alembic import op
 
 # revision identifiers, used by Alembic.
@@ -22,7 +21,7 @@ def upgrade() -> None:
     op.create_table(
         "usergroup",
         sa.Column("id", sa.Integer(), nullable=False),
-        sa.Column("name", sqlmodel.sql.sqltypes.AutoString(), nullable=False),
+        sa.Column("name", sa.String(), nullable=False),
         sa.Column(
             "timestamp_created", sa.DateTime(timezone=True), nullable=False
         ),